+++ sei-db/db_engine/litt/cli/ls_test.go
@@ -28,7 +28,7 @@ func TestLs(t *testing.T) {
 	require.NoError(t, err)
 
 	// Make it so that we have at least as many shards as roots.
-	config.ShardingFactor = rootCount * rand.Uint32Range(1, 4)
+	config.ShardingFactor = uint8(rootCount * rand.Uint32Range(1, 4))
 
 	// Settings that should be enabled for LittDB unit tests.
 	config.DoubleWriteProtection = true
